FIXME: Check article parent for save permission.
This interim "fix" has side effects, since the permission is only checking the current defined parent.  BSE actually allows the selection of a different parent at the time of creation, however if a different parent is selected, the access permission cannot determine access privilege to ti until the form is submitted, which may result in an access error.