Fix broken metadata macro.
This was broken since the introduction of the is_disabled() field check as used by the article editor (edit_1).

Since the metadata macro can be used elsewhere it needs to handle disabled fields accordingly.

Set the “object_type” to “file” for correct handling on the file manager.  The default is “article”.