Support for custom field checkbox and radio inputs.
Conditionally include a hidden field with save_name matching the original field name and an empty value to support saving of unchecked checkbox and radio inputs.

This save_name hidden field should only be included if the checkbox or radio input currently has a value and it MUST have an empty value.